requirements.txt
wxPython
cryptography
python-dotenv